It is a direct graph.
Use Map<Integer, Map<Integer, Integer>> to store the source node,
target node and the distance between them.
Offer the node K to a PriorityQueue.
Then keep getting the closest nodes to the current node and calculate
the distance from the source (K) to this node (absolute distance). Use
a Map to store the shortest absolute distance of each node.
Return the node with the largest absolute distance.