Generally, there are following steps:
create a hashmap for each character in t and count their frequency in t
as the value of hashmap.
Find the first window in S that contains T. But how? there the author
uses the count.
Checking from the leftmost index of the window and to see if it belongs
to t. The reason we do so is that we want to shrink the size of the
window.
3-1) If the character at leftmost index does not belong to t, we can
directly remove this leftmost value and update our window(its minLeft
and minLen value)
3-2) If the character indeed exists in t, we still remove it, but in
the next step, we will increase the right pointer and expect the
removed character. If find so, repeat step 3.